python-similarity/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0
scipy>=1.13.0
pydantic>=2.7.0
python-dotenv>=1.0.0
certifi>=2024.0.0
